Implement Forgot Password window
That took an interesting route. I guess since now I'll make other windows just like Login or Registration: They will be so called sub frames for the main frame in MainApp, which located in main.py
I kinda like how it looks. Maybe the password restoration requires some more corrections, but I'll deal with it on Saturday. It supposed to be the day where all requirements will be achieved, or how do I put it